requirements.txt
requests
gradio
openai-whisper @ git+https://github.com/openai/whisper.git
tqdm
yt_dlp